Refactor bfacf move funtions into their own files
